# docs/swagger.yaml
basePath: /v1
definitions:
  github_com_DIMO-Network_trips-api_internal_helper.PointTime:
    properties:
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      time:
        example: "2023-05-04T09:00:00Z"
        type: string
    type: object
  github_com_DIMO-Network_trips-api_internal_helper.TripDetails:
    properties:
      end:
        $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.TripEvent'
      id:
        example: 2Y83IHPItgk0uHD7hybGnA776Bo
        type: string
      start:
        $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.TripEvent'
    type: object
  github_com_DIMO-Network_trips-api_internal_helper.TripEvent:
    properties:
      actual:
        $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.PointTime'
      estimate:
        $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.PointTime'
    type: object
  github_com_DIMO-Network_trips-api_internal_helper.VehicleTrips:
    properties:
      currentPage:
        example: 1
        type: integer
      totalPages:
        example: 1
        type: integer
      trips:
        items:
          $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.TripDetails'
        type: array
    type: object
info:
  contact: {}
  description: segments
  title: DIMO Segment API
  version: "1.0"
paths:
  /vehicle/{tokenId}/trips:
    get:
      description: Lists vehicle trips.
      parameters:
      - description: Vehicle token id
        in: path
        name: tokenId
        required: true
        type: integer
      - description: Page of trips to retrieve. Defaults to 1.
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DIMO-Network_trips-api_internal_helper.VehicleTrips'
      security:
      - BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"